first visualization, cleaning steps, & author names added
