+++ docs/search.json
@@ -18,14 +18,14 @@
     "href": "index.html#data",
     "title": "Final Project: [YOUR TITLE HERE]",
     "section": "Data",
-    "text": "Data\nDescribe your data source(s), including where it came from, the key variables, and any pre-processing steps.\n\n\nCode\n# Load libraries and data\nlibrary(tidyverse)\n\ndata &lt;- read_csv(\"data/my_data.csv\")\nglimpse(data)"
+    "text": "Data\nDescribe your data source(s), including where it came from, the key variables, and any pre-processing steps.\n\n\nCode\n# Load libraries and data\nlibrary(tidyverse)\nlibrary(dplyr)\nlibrary(haven)\n\nraw_data &lt;- read_dta(\"data/Guttmacher 2009 Survey of contraceptive knowledge (Fog Zone).dta\")\n\nclean_fogzone &lt;- raw_data %&gt;% \n  select(gender,age4grps,race,riskr,abstonly,protect,curmethrr,pregknowr,totknowr,nummthdsr,wprotect,mprotect) %&gt;% \n  rename(\n    Current_most_effective_method = curmethrr,\n    Pregnancy_knowledge = pregknowr,\n    BC_method_knowledge = totknowr,\n    BC_methods_heard_about = nummthdsr,\n    Age = age4grps,\n    Women_protection_status = wprotect,\n    Men_protection_status = mprotect,\n    Protection_status = protect\n  ) %&gt;% \n  mutate(\n    gender = case_when(\n      gender == 1 ~ \"Female\",\n      gender == 2 ~ \"Male\",\n      TRUE ~ NA_character_\n    ),\n    race = case_when(\n      race == 1 ~ \"Hispanic\",\n      race == 2 ~ \"Non-Hispanic White\",\n      race == 3 ~ \"Non-hispanic Black\",\n      race == 4 ~ \"Asian\",\n      race == 5 ~ \"Other\",\n      TRUE ~ NA_character_\n    ),\n    Age = case_when(\n      Age == 1 ~ \"18-19\",\n      Age == 2 ~ \"20-21\",\n      Age == 3 ~ \"22-24\",\n      Age == 4 ~ \"25-29\",\n      TRUE ~ NA_character_\n    ),\n    riskr = case_when(\n      riskr == 1 ~ \"No\",\n      riskr == 2 ~ \"Yes\",\n      TRUE ~ NA_character_\n    ), \n    abstonly = case_when(\n      abstonly == 1 ~ \"Abstinence Only\",\n      abstonly == 2 ~ \"Abstinence Focused\",\n      abstonly == 3 ~ \"Contraceptive Focused\",\n      is.na(abstonly) ~ \"No sex ed/ don't know\",\n      TRUE ~ NA_character_\n    ), \n    Protection_status = case_when(\n      Protection_status == 1 ~ \"Well protected\",\n      Protection_status == 2 ~ \"Unprotected\",\n      Protection_status == 3 ~ \"Uncertain status\",\n      Protection_status == 7 ~ \"No Sex\",\n      TRUE ~ NA_character_\n    ),\n    Women_protection_status = case_when(\n      Women_protection_status == 1 ~ \"perfect pill use\",\n      Women_protection_status == 2 ~ \"perfect condom use\",\n      Women_protection_status == 3 ~ \"perfect depo use\",\n      Women_protection_status == 4 ~ \"long-acting method\",\n      Women_protection_status == 5 ~ \"skipped pills\",\n      Women_protection_status == 6 ~ \"missed condom use\",\n      Women_protection_status == 7 ~ \"depo late\",\n      Women_protection_status == 8 ~ \"other method\",\n      Women_protection_status == 9 ~ \"no method - likely at risk\",\n      Women_protection_status == 17 ~ \"no method - unkown risk\",\n      Women_protection_status == 19 ~ \"no sex ever/last 12 mo\",\n      TRUE ~ NA_character_\n    ),\n    Men_protection_status = case_when(\n      Men_protection_status == 1 ~ \"use horm/larc\",\n      Men_protection_status == 2 ~ \"perfect condom use\",\n      Men_protection_status == 3 ~ \"inconsistent condom use\",\n      Men_protection_status == 4 ~ \"withdrawl\",\n      Men_protection_status == 5 ~ \"other/unknown\",\n      Men_protection_status == 6 ~ \"no method - risk likely\",\n      Men_protection_status == 7 ~ \"no method - unknown risk\",\n      Men_protection_status == 8 ~ \"no sex ever/ last 12 mo\",\n      TRUE ~ NA_character_\n    ), \n    Current_most_effective_method = case_when(\n      Current_most_effective_method == 1 ~ \"long-acting alone\",\n      Current_most_effective_method == 2 ~ \"long-acting + condom\",\n      Current_most_effective_method == 3 ~ \"hormonal alone\",\n      Current_most_effective_method == 4 ~ \"hormonal + condoms\",\n      Current_most_effective_method == 5 ~ \"condoms\",\n      Current_most_effective_method == 6 ~ \"withdrawl/NFP\",\n      Current_most_effective_method == 7 ~ \"other/unknown\",\n      Current_most_effective_method == 17 ~ \"no method - risk unknown\",\n      Current_most_effective_method == 18 ~ \"no method - likely at risk\",\n      Current_most_effective_method == 19 ~ \"no sex ever/ last 12 mo\",\n      TRUE ~ NA_character_\n    ),\n    Pregnancy_knowledge = case_when(\n      Pregnancy_knowledge == 1 ~ \"0-3 correct\",\n      Pregnancy_knowledge == 4 ~ \"4 correct\",\n      Pregnancy_knowledge == 5 ~ \"5 correct\",\n      Pregnancy_knowledge == 6 ~ \"6 correct\",\n      TRUE ~ NA_character_\n    ),\n    BC_method_knowledge = case_when(\n      BC_method_knowledge == 1 ~ \"0-7 correct\",\n      BC_method_knowledge == 2 ~ \"8-11 correct\",\n      BC_method_knowledge == 3 ~ \"12-15 correct\",\n      BC_method_knowledge == 4 ~ \"16-18 correct\",\n      BC_method_knowledge == 5 ~ \"19+ correct\",\n      TRUE ~ NA_character_\n    ),\n    BC_methods_heard_about = case_when(\n      BC_methods_heard_about == 1 ~ \"0-6 methods\",\n      BC_methods_heard_about == 2 ~ \"7-10 methods\",\n      BC_methods_heard_about== 3 ~ \"11-13 methods\",\n      BC_methods_heard_about == 4 ~ \"14-16 methods\",\n      TRUE ~ NA_character_\n    )\n  )"
   },
   {
-    "objectID": "index.html#visualizatiob-1-title-here",
-    "href": "index.html#visualizatiob-1-title-here",
+    "objectID": "index.html#visualization-1-sexual-health-education-and-risk-of-unintended-pregnancy",
+    "href": "index.html#visualization-1-sexual-health-education-and-risk-of-unintended-pregnancy",
     "title": "Final Project: [YOUR TITLE HERE]",
-    "section": "Visualizatiob 1: [Title Here]",
-    "text": "Visualizatiob 1: [Title Here]\nDescribe what this plot shows, and why it’s important.\n\n\nCode\n# your code here"
+    "section": "Visualization 1: [Sexual Health Education and Risk of Unintended Pregnancy]",
+    "text": "Visualization 1: [Sexual Health Education and Risk of Unintended Pregnancy]\nDescribe what this plot shows, and why it’s important.\n\n\nCode\n# your code here\nlibrary(ggplot2)\nlibrary(viridis)\nlibrary(dplyr)\n\nclean_fogzone &lt;- readRDS(\"data/clean_fogzone.rds\")\n\n\nclean_fogzone_factors &lt;- clean_fogzone %&gt;%\n  mutate(\n    abstonly = factor(\n      abstonly,\n      levels = c(\"No sex ed/ don't know\", \"Abstinence Only\", \"Abstinence Focused\", \"Contraceptive Focused\")\n    ),\n    riskr = factor(riskr, levels = c(\"No\", \"Yes\"))\n  )\n\nggplot(clean_fogzone_factors, aes(x = abstonly, fill = riskr)) +\n  geom_bar(position = position_dodge(width = 0.9)) +\n  geom_text(stat = \"count\", aes(label = ..count..), \n            position = position_dodge(width = 0.9), \n            vjust = -0.4, size = 3.5) +\n  scale_fill_viridis_d(\n    name = \"At Risk?\",\n    labels = c(\"No\", \"Yes\"),\n    option = \"D\"\n  ) +\nlabs(\n  title = \"Sexual Health Education and Risk of Unintended Pregnancy\",\n  x = \"Type of Sexual Health Education Recived\",\n  y = \"Number of Participants\"\n  ) + \ntheme_minimal(base_size = 14) +\ntheme(\n  axis.text.x = element_text(angle = 20, hjust = 1, face = \"bold\", size = 10),\n  axis.text.y = element_text(face = \"bold\"),\n  plot.title = element_text(face = \"bold\"),\n  legend.position = \"top\"\n)"
   },
   {
     "objectID": "index.html#visualization-2-title-here",
