+++ docs/site_libs/quarto-html/quarto-syntax-highlighting.css
@@ -36,8 +36,11 @@
   --quarto-font-monospace: SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace;
 }
 
-/* syntax highlight based on Pandoc's rules */
 pre > code.sourceCode > span {
+  color: #003B4F;
+}
+
+code span {
   color: #003B4F;
 }
 
@@ -50,187 +53,153 @@ div.sourceCode pre.sourceCode {
   color: #003B4F;
 }
 
-/* Normal */
-code span {
+code span.ot {
   color: #003B4F;
-}
-
-/* Alert */
-code span.al {
-  color: #AD0000;
   font-style: inherit;
 }
 
-/* Annotation */
-code span.an {
-  color: #5E5E5E;
-  font-style: inherit;
-}
-
-/* Attribute */
 code span.at {
   color: #657422;
   font-style: inherit;
 }
 
-/* BaseN */
-code span.bn {
-  color: #AD0000;
+code span.ss {
+  color: #20794D;
   font-style: inherit;
 }
 
-/* BuiltIn */
-code span.bu {
+code span.an {
+  color: #5E5E5E;
   font-style: inherit;
 }
 
-/* ControlFlow */
+code span.fu {
+  color: #4758AB;
+  font-style: inherit;
+}
+
+code span.st {
+  color: #20794D;
+  font-style: inherit;
+}
+
 code span.cf {
   color: #003B4F;
   font-weight: bold;
   font-style: inherit;
 }
 
-/* Char */
-code span.ch {
-  color: #20794D;
-  font-style: inherit;
-}
-
-/* Constant */
-code span.cn {
-  color: #8f5902;
-  font-style: inherit;
-}
-
-/* Comment */
-code span.co {
+code span.op {
   color: #5E5E5E;
   font-style: inherit;
 }
 
-/* CommentVar */
-code span.cv {
-  color: #5E5E5E;
-  font-style: italic;
-}
-
-/* Documentation */
-code span.do {
-  color: #5E5E5E;
-  font-style: italic;
-}
-
-/* DataType */
-code span.dt {
-  color: #AD0000;
-  font-style: inherit;
-}
-
-/* DecVal */
-code span.dv {
-  color: #AD0000;
-  font-style: inherit;
-}
-
-/* Error */
 code span.er {
   color: #AD0000;
   font-style: inherit;
 }
 
-/* Extension */
-code span.ex {
-  font-style: inherit;
-}
-
-/* Float */
-code span.fl {
+code span.bn {
   color: #AD0000;
   font-style: inherit;
 }
 
-/* Function */
-code span.fu {
-  color: #4758AB;
+code span.al {
+  color: #AD0000;
   font-style: inherit;
 }
 
-/* Import */
-code span.im {
-  color: #00769E;
+code span.va {
+  color: #111111;
   font-style: inherit;
 }
 
-/* Information */
+code span.bu {
+  font-style: inherit;
+}
+
+code span.ex {
+  font-style: inherit;
+}
+
+code span.pp {
+  color: #AD0000;
+  font-style: inherit;
+}
+
 code span.in {
   color: #5E5E5E;
   font-style: inherit;
 }
 
-/* Keyword */
+code span.vs {
+  color: #20794D;
+  font-style: inherit;
+}
+
+code span.wa {
+  color: #5E5E5E;
+  font-style: italic;
+}
+
+code span.do {
+  color: #5E5E5E;
+  font-style: italic;
+}
+
+code span.im {
+  color: #00769E;
+  font-style: inherit;
+}
+
+code span.ch {
+  color: #20794D;
+  font-style: inherit;
+}
+
+code span.dt {
+  color: #AD0000;
+  font-style: inherit;
+}
+
+code span.fl {
+  color: #AD0000;
+  font-style: inherit;
+}
+
+code span.co {
+  color: #5E5E5E;
+  font-style: inherit;
+}
+
+code span.cv {
+  color: #5E5E5E;
+  font-style: italic;
+}
+
+code span.cn {
+  color: #8f5902;
+  font-style: inherit;
+}
+
+code span.sc {
+  color: #5E5E5E;
+  font-style: inherit;
+}
+
+code span.dv {
+  color: #AD0000;
+  font-style: inherit;
+}
+
 code span.kw {
   color: #003B4F;
   font-weight: bold;
   font-style: inherit;
 }
 
-/* Operator */
-code span.op {
-  color: #5E5E5E;
-  font-style: inherit;
-}
-
-/* Other */
-code span.ot {
-  color: #003B4F;
-  font-style: inherit;
-}
-
-/* Preprocessor */
-code span.pp {
-  color: #AD0000;
-  font-style: inherit;
-}
-
-/* SpecialChar */
-code span.sc {
-  color: #5E5E5E;
-  font-style: inherit;
-}
-
-/* SpecialString */
-code span.ss {
-  color: #20794D;
-  font-style: inherit;
-}
-
-/* String */
-code span.st {
-  color: #20794D;
-  font-style: inherit;
-}
-
-/* Variable */
-code span.va {
-  color: #111111;
-  font-style: inherit;
-}
-
-/* VerbatimString */
-code span.vs {
-  color: #20794D;
-  font-style: inherit;
-}
-
-/* Warning */
-code span.wa {
-  color: #5E5E5E;
-  font-style: italic;
-}
-
 .prevent-inlining {
   content: "</";
 }
 
-/*# sourceMappingURL=32badd10cd2e2abee3c1d930e287fa2a.css.map */
+/*# sourceMappingURL=debc5d5d77c3f9108843748ff7464032.css.map */
